requirements.txt
backports.tarfile==1.2.0
base58==2.1.1
bidict==0.23.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
keyring==25.7.0
mnemonic==0.21
more-itertools==10.8.0
pycparser==2.23
PyNaCl==1.6.2
python-engineio==4.13.0
python-socketio==5.16.0
requests==2.32.5
simple-websocket==1.1.0
urllib3==2.6.2
websocket-client==1.9.0
wsproto==1.2.0
zipp==3.23.0
